requirements.txt
transformers
torchvision
Pillow
torch
streamlit
streamlit_webrtc
sentence-transformers
beautifulsoup4
langchain
psycopg2
numpy
psycopg2-binary
SQLAlchemy
cmake
clang
torchvision 
torchaudio
requests
plotly_express 
clean-text 
ipython-sql 
pandas 
scikit-learn 
seaborn 
matplotlib 
sqlalchemy
sql_magic 
wordcloud 
langdetect
opencv-python
pgvector
